move testing folder
