+++ src/tests/UserTexts.test.tsx
@@ -1,6 +1,6 @@
 import '@testing-library/jest-dom';
-import loginServices from '../../services/login';
-import textsServices from '../../services/texts';
+import loginServices from '../services/login';
+import textsServices from '../services/texts';
 
 describe('Tests user ability to interact with texts', () => {
   test('Renders user text by language id', async () => {
